добавляет комментарии
